API Refactor transformations into `transforms` folder
