+++ client/src/transforms/TreeDropdownField/TreeDropdownField.md
@@ -1,0 +1,6 @@
+## TreeDropdownField transformations
+
+Provides injector transformations for the following TreeDropdownField components:
+
+ * moveTreeDropdownField: When selecting a "move to" location in a tree, this transformation will disable
+   all nodes which are also in the "from", thereby preventing a parent being its own child.
